+++ src/_tests/tests_run_agent_script.rs
@@ -2,7 +2,7 @@ type Result<T> = core::result::Result<T, Box<dyn std::error::Error>>; // For tes
 
 use super::*;
 use crate::_test_support::{assert_contains, load_inline_agent, load_test_agent, run_test_agent_with_input};
-use crate::types::FileMeta;
+use crate::types::FileInfo;
 use simple_fs::SPath;
 
 #[tokio::test]
@@ -86,7 +86,7 @@ return {
 
 	// -- Execute
 	let on_path = SPath::new("./some-random/file.txt");
-	let path_ref = FileMeta::new(on_path, false);
+	let path_ref = FileInfo::new(on_path, false);
 	let inputs = vec![serde_json::to_value(path_ref)?];
 
 	let res = run_command_agent(&runtime, agent, Some(inputs), &RunBaseOptions::default(), true).await?;
@@ -111,7 +111,7 @@ async fn test_run_agent_script_before_all_response_simple() -> Result<()> {
 
 	// -- Execute
 	let on_path = SPath::new("./some-random/file.txt");
-	let path_ref = FileMeta::new(on_path, false);
+	let path_ref = FileInfo::new(on_path, false);
 	let inputs = vec![serde_json::to_value(path_ref)?];
 
 	let res = run_command_agent(&runtime, agent, Some(inputs), &RunBaseOptions::default(), true).await?;