+++ src/types/mod.rs
@@ -1,7 +1,7 @@
 // region:    --- Modules
 
 mod dest_options;
-mod file_meta;
+mod file_info;
 mod file_record;
 mod md_block;
 mod md_heading;
@@ -11,7 +11,7 @@ mod pack_ref;
 mod tag_block;
 
 pub use dest_options::*;
-pub use file_meta::*;
+pub use file_info::*;
 pub use file_record::*;
 pub use md_block::*;
 pub use md_heading::*;